功能完善：建议把dash_style属性放到Control中 #94 （1）画笔的宽度改为float类型，避免在DPI缩放后线条的宽度出现明显失真的现象（比如缩放比为135%，175%等情况下，整数时都会出现四舍五入的现象） （2）控件的边线宽度改为float类型，避免避免在DPI缩放后线条的宽度出现明显失真的现象 （3）完善ColorPicker的选择框功能，避免选择框被部分覆盖而影响美观度。

+++ duilib/Control/ColorPickerStatardGray.h
@@ -50,7 +50,7 @@ public:
     * @param [in] brushColor 画刷的颜色，如果为0，则填充颜色
     */
     bool DrawRegularHexagon(IRender* pRender, const UiPointF& centerPt, int32_t radius,
-                            const UiColor& penColor, int32_t penWidth, const UiColor& brushColor);
+                            const UiColor& penColor, float penWidth, const UiColor& brushColor);
 
     /** 鼠标移动
     */
